Added error checking for unknown types
Also:
* refactored tests
* setup linting
