+++ .eslintrc.json
@@ -1,0 +1,14 @@
+{
+  "env": {
+    "node": true,
+    "mocha": true
+  },
+  "extends": "eslint:recommended",
+  "globals": {
+  },
+  "parserOptions": {
+    "ecmaVersion": 8
+  },
+  "rules": {
+  }
+}
