+++ .travis.yml
@@ -1,5 +1,6 @@
 sudo: false
 language: node_js
+script: npm run ci
 
 cache:
   directories:
