+++ package.json
@@ -1,10 +1,11 @@
 {
   "name": "jsonschema-bigquery",
-  "version": "2.0.0",
+  "version": "2.0.1",
   "description": "Convert JSON schema to Google BigQuery schema",
   "main": "src/converter.js",
   "scripts": {
-    "test": "mocha test",
+    "ci": "npm run lint && npm test",
+    "test": "mocha test/unit test/integration",
     "lint": "eslint ."
   },
   "bin": {
@@ -34,13 +35,7 @@
   },
   "homepage": "https://github.com/thedumbterminal/jsonschema-bigquery#readme",
   "devDependencies": {
-    "eslint": "^5.6.1",
-    "eslint-config-standard": "^12.0.0",
-    "eslint-plugin-import": "^2.14.0",
-    "eslint-plugin-mocha": "^5.2.0",
-    "eslint-plugin-node": "^7.0.1",
-    "eslint-plugin-promise": "^4.0.1",
-    "eslint-plugin-standard": "^4.0.0",
+    "eslint": "^5.16.0",
     "mocha": "^5.2.0"
   },
   "dependencies": {
