+++ test/integration/samples/allOf/expected.json
