+++ test/integration/samples/allOf/input.json
