+++ test/integration/samples/allOf_nested/expected.json
