+++ test/integration/samples/allOf_nested/input.json
