+++ test/integration/samples/anyOfMultipleTypes/expected.json
