+++ test/integration/samples/anyOfMultipleTypes/input.json
