+++ test/integration/samples/nestedDescription/expected.json
