+++ test/integration/samples/nestedDescription/input.json
