+++ test/integration/samples/oneof/expected.json
