+++ test/integration/samples/oneof/input.json
