+++ test/integration/samples/timestamp/expected.json
