+++ test/integration/samples/timestamp/input.json
