+++ test/integration/samples/typeArray/expected.json
