+++ test/integration/samples/typeArray/input.json
